UI nearing completion
Someone that likes the idea of being a front end engineer can fix the
colors of the UI. Also, I still need to implement the TLM and CMD queues
and then the pilotUpdate() function will handle the queues and from
there the info will go out/come in over the sockets. We're getting close
to completing this thing. Tomorrow I will implement both sides of the
TLM queue. That will be a good test of the socket stuff.